face-recognition-multi-camera/requirements.txt
tensorflow
scipy
pandas
scikit-learn
opencv-python
h5py
cmake
dlib
numpy
boto3
jsonpickle
matplotlib
Pillow
requests
psutil
torch
torchvision
mysql-connector